Fixed argument handling that was broken by 51c205652bafceff47cb66c214f5663e5f533aef
When a source for a parameter is specified as a string the whole source should be passed. This is extremely useful when you want to pass all get parameters as a parameter. The 51c205652bafceff47cb66c214f5663e5f533aef commit just turned the 'string as source' into syntactic sugar for: get the value that has the same as the argument from the source named in the string.